fix: extract out JSX rendering routes to new file
