+++ src/utils/reactdemo.tsx
@@ -15,7 +15,7 @@ const Button: FC<{}> = ({ children }) => {
   );
 };
 
-const ResetPasswordPage: FC<{ error?: string }> = ({ error }) => {
+export const ResetPasswordPage: FC<{ error?: string }> = ({ error }) => {
   return (
     <Layout title="Reset Password">
       <div class="mb-8 text-2xl font-medium">Reset password</div>
@@ -35,7 +35,7 @@ const ResetPasswordPage: FC<{ error?: string }> = ({ error }) => {
   );
 };
 
-export async function renderReactThing(
+async function renderReactThing(
   ctx: Context<{ Bindings: Env; Variables: Var }>,
   error?: string,
   status?: number,
